+++ src/Subscribers/ToolSubscribers.cpp
@@ -84,7 +84,7 @@ bool ToolSubscribers::toolStandbyTemp(Comm::JsonDecoder* decoder, const int32_t&
 bool ToolSubscribers::toolSpindle(Comm::JsonDecoder* decoder, const int32_t& data, const size_t indices[])
 {
 	UNUSED(decoder);
-	if (!OM::UpdateToolSpindle(indices[0], data))
+	if (!OM::UpdateToolSpindle(indices[0], static_cast<int8_t>(data)))
 	{
 		LOG_ERROR("Failed to update tool {:d} spindle to {:d}", indices[0], data);
 		return false;
@@ -95,7 +95,7 @@ bool ToolSubscribers::toolSpindle(Comm::JsonDecoder* decoder, const int32_t& dat
 bool ToolSubscribers::toolSpindleRpm(Comm::JsonDecoder* decoder, const int32_t& data, const size_t indices[])
 {
 	UNUSED(decoder);
-	if (!OM::UpdateToolSpindleRpm(indices[0], data))
+	if (!OM::UpdateToolSpindleRpm(indices[0], static_cast<int8_t>(data)))
 	{
 		LOG_ERROR("Failed to update tool {:d} spindleRpm to {:d}", indices[0], data);
 		return false;
@@ -149,7 +149,7 @@ bool ToolSubscribers::toolArrayEnd(Comm::JsonDecoder* decoder, const size_t indi
 bool ToolSubscribers::toolHeaterArrayEnd(Comm::JsonDecoder* decoder, const size_t indices[])
 {
 	UNUSED(decoder);
-	if (OM::RemoveToolHeaters(indices[0], static_cast<uint8_t>(indices[1])))
+	if (OM::RemoveToolHeaters(indices[0], indices[1]))
 	{
 	}
 	Model::get().post<EventType::ToolHeaterData>(indices[0]);
@@ -159,7 +159,7 @@ bool ToolSubscribers::toolHeaterArrayEnd(Comm::JsonDecoder* decoder, const size_
 bool ToolSubscribers::toolExtruderArrayEnd(Comm::JsonDecoder* decoder, const size_t indices[])
 {
 	UNUSED(decoder);
-	if (OM::RemoveToolExtruders(indices[0], static_cast<uint8_t>(indices[1])))
+	if (OM::RemoveToolExtruders(indices[0], indices[1]))
 	{
 	}
 	return true;
@@ -168,7 +168,7 @@ bool ToolSubscribers::toolExtruderArrayEnd(Comm::JsonDecoder* decoder, const siz
 bool ToolSubscribers::toolFanArrayEnd(Comm::JsonDecoder* decoder, const size_t indices[])
 {
 	UNUSED(decoder);
-	if (OM::RemoveToolFans(indices[0], static_cast<uint8_t>(indices[1])))
+	if (OM::RemoveToolFans(indices[0], indices[1]))
 	{
 	}
 	return true;
